[`refurb`] Minor nits regarding `for-loop-writes` and `for-loop-set-mutations`
