+++ crates/ruff_linter/src/rules/refurb/rules/mod.rs
@@ -42,6 +42,7 @@ mod for_loop_writes;
 mod fstring_number_format;
 mod hardcoded_string_charset;
 mod hashlib_digest_hex;
+mod helpers;
 mod if_exp_instead_of_or_operator;
 mod if_expr_min_max;
 mod implicit_cwd;
